timerUnit.h und timerUnit.c: timer_getCounter() refaktorisiert: neues Rückgabetyp und mit neuen Namen der Globalvariable angepasst

+++ 01 Hardware/01 Lidar/03 Software/01 source/Testen mit USVData/Testen mit USVData/timerUnit.h
@@ -41,7 +41,7 @@ typedef struct{
 extern uint8_t timerInit(uint8_t rezConfig, uint8_t resolution);
 extern void timer_setState(uint8_t state);
 extern void timer_setCounter(uint32_t value);
-extern const int32_t timer_getCounter();
+extern const int16_t timer_getCounter();
 extern void timer_stopWatch(uint16_t val);
 
 #endif /* TIMERUNIT_H_ */
